# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CPay API
  version: v1
servers:
  - url: http://localhost:8080
paths:
  /v1/auth/login:
    post:
      summary: Login and get JWT pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Success
...
  /v1/payment_links:
    post:
      summary: Create payment link
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '201':
          description: Created
    get:
      summary: List payment links
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Success
  /v1/payment_links/{id}:
    get:
      summary: Get payment link
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Success
        '404':
          description: Not found
    patch:
      summary: Update payment link
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                pricing_mode:
                  type: string
                  enum: [fixed, open]
                amount:
                  type: number
                  nullable: true
                currency:
                  type: string
                metadata:
                  type: object
      responses:
        '200':
          description: Updated
        '404':
          description: Not found
  /v1/payment_links/{id}/archive:
    post:
      summary: Archive payment link
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Archived
        '404':
          description: Not found
  /v1/payment_links/{id}/sessions:
    post:
      summary: Create checkout session from payment link
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '201':
          description: Created
  /v1/public/payment_links/{code}:
    get:
      summary: Get public payment link
      parameters:
        - in: path
          name: code
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  code: { type: string }
                  title: { type: string }
                  description: { type: string, nullable: true }
                  image_url: { type: string, nullable: true }
                  pricing_mode: { type: string, enum: [fixed, open] }
                  amount: { type: number, nullable: true }
                  currency: { type: string }
                  cta_text: { type: string }
                  allowed_tokens:
                    type: array
                    items:
                      type: object
                      properties:
                        chain: { type: string }
                        symbol: { type: string }
                        address: { type: string }
                        decimals: { type: integer }
                  customer_fields:
                    type: array
                    items: { type: string }
                  custom_fields:
                    type: array
                    items: { type: object }
                  after_payment:
                    type: object
                    properties:
                      type: { type: string }
                      success_message: { type: string, nullable: true }
                      redirect_url: { type: string, nullable: true }
                  options:
                    type: object
                  product:
                    type: object
                    nullable: true
        '404':
          description: Not found
...
  /v1/checkout/{session_id}/confirm:
    post:
      summary: Confirm observed chain transaction for checkout
      parameters:
        - in: path
          name: session_id
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Updated
...
  /v1/webhook_endpoints:
    post:
      summary: Register merchant webhook endpoint
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '201':
          description: Created
  /v1/subscriptions:
    post:
      summary: Create prepaid-vault subscription
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '201':
          description: Created
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key